+++ etna/transforms/outliers/point_outliers.py
@@ -122,7 +122,13 @@ class DensityOutliersTransform(OutliersTransform):
 
 
 class PredictionIntervalOutliersTransform(OutliersTransform):
-    """Transform that uses :py:func:`~etna.analysis.outliers.prediction_interval_outliers.get_anomalies_prediction_interval` to find anomalies in data."""
+    """Transform that uses :py:func:`~etna.analysis.outliers.prediction_interval_outliers.get_anomalies_prediction_interval` to find anomalies in data.
+
+    Warning
+    -------
+    Currently, pickle is used in ``save`` and  ``load`` methods.
+    It can work unreliably according to `documentation <https://facebook.github.io/prophet/docs/additional_topics.html>`_.
+    """
 
     def __init__(
         self,